backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
httpx[socks]
duckduckgo-search
